feat: report optimization (table - search&filters - pagination)
